completed functionality to update a patient

+++ UserControls/PatientsUserControls.cs
@@ -14,18 +14,20 @@ namespace SmartClinic.UserControls
 {
     public partial class PatientsUserControls : UserControl
     {
+        private int rowIndexForDataGridView;
         private readonly PatientController patientController;
         public PatientsUserControls()
         {
             InitializeComponent();
             this.patientController = new PatientController();
+            this.rowIndexForDataGridView = 0;
         }
 
         private void SearchButton_Click(object sender, EventArgs e)
         {
             var firstName = firstNameTextBox.Text;
             var lastName = lastNameTextBox.Text;
-      
+
             DateTime? dateOfBirth = null;
             if (dobSearchPicker.Checked)
             {
@@ -47,32 +49,36 @@ namespace SmartClinic.UserControls
 
         private void SelectRow_DoubleClick(object sender, DataGridViewCellEventArgs e)
         {
-            Console.WriteLine("it worked");
+            this.viewOrEditPatientDetailsButton.PerformClick();
         }
 
         private void DeletePatientButton_Click(object sender, EventArgs e)
         {
-            int patientID = Int32.Parse(patientDataGridView.Rows[patientDataGridView.CurrentCell.RowIndex].Cells[0].Value.ToString());
-            
-
-            if (this.patientController.PatientHasNoAppointments(patientID))
+            if (this.patientDataGridView.Rows.Count == 0)
             {
-                DialogResult dialogResultVerifyClose = MessageBox.Show("Are you sure you want to delete this patient?\n" +
+                MessageBox.Show("There is no patient to delete.", "", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                return;
+            }
+
+            int patientID = Int32.Parse(patientDataGridView.Rows[patientDataGridView.CurrentCell.RowIndex].Cells[0].Value.ToString());
+
+            DialogResult dialogResultVerifyClose = MessageBox.Show("Are you sure you want to delete this patient?\n" +
                "This cannot be undone.", "Warning", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
-                
-                if (dialogResultVerifyClose == DialogResult.Yes)
+
+            if (dialogResultVerifyClose == DialogResult.Yes)
+            {
+                if (this.patientController.PatientHasNoAppointments(patientID))
                 {
                     //need try-catch
                     this.patientController.DeletePatient(patientID);
                     this.searchButton.PerformClick();
                 }
+                else
+                {
+                    MessageBox.Show("This patient has associated appointments\nand cannot be deleted.", "", MessageBoxButtons.OK, MessageBoxIcon.Error);
+
+                }
             }
-            else
-            {
-                MessageBox.Show("This patient has associated appointments\nand cannot be deleted.", "", MessageBoxButtons.OK, MessageBoxIcon.Error);
-                
-            }
-                
         }
 
         private void ClearButton_Click(object sender, EventArgs e)
@@ -88,19 +94,43 @@ namespace SmartClinic.UserControls
         private void SearchTermsValueChanged(object sender, EventArgs e)
         {
             this.patientDataGridView.DataSource = null;
+            this.searchMessageLabel.Text = "";
         }
 
         private void ViewOrEditPatientDetailsButton_Click(object sender, EventArgs e)
         {
-            int patientID = Int32.Parse(patientDataGridView.Rows[patientDataGridView.CurrentCell.RowIndex].Cells[0].Value.ToString());
+            if (this.patientDataGridView.Rows.Count == 0)
+            {
+                MessageBox.Show("There is no patient to select.", "", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                return;
+            }
+            int patientID = Int32.Parse(this.patientDataGridView.Rows[patientDataGridView.CurrentCell.RowIndex].Cells[0].Value.ToString());
             using (Form updateForm = new UpdatePatient(patientID))
             {
                 DialogResult result = updateForm.ShowDialog();
+                this.searchButton.PerformClick();
 
-                if (result == DialogResult.OK)
-                {
-                  
-                }
+            }
+        }
+
+        private void PatientsTable_RighClick(object sender, DataGridViewCellMouseEventArgs e)
+        {
+            
+            if (e.Button == MouseButtons.Right)
+            {
+                this.patientDataGridView.Rows[e.RowIndex].Selected = true;
+                this.rowIndexForDataGridView = e.RowIndex;
+                this.patientDataGridView.CurrentCell = this.patientDataGridView.Rows[e.RowIndex].Cells[1];
+                this.RightClickMenu.Show(this.patientDataGridView, e.Location);
+                RightClickMenu.Show(Cursor.Position);
+            }
+        }
+
+        private void ContextMenuForDataGridView_Click(object sender, MouseEventArgs e)
+        {
+            if (!this.patientDataGridView.Rows[this.rowIndexForDataGridView].IsNewRow)
+            {
+                this.deletePatientButton.PerformClick();
             }
         }
     }
